replaced doubles with single-precision floats. Was it worth it?

+++ Flosion/NumberInput.cpp
@@ -90,7 +90,7 @@ namespace musical {
 	// or from sound processing objects, which shall be found as parents of some state
 	// in the state chain (accessed through the line of parents in the given state)
 	// this state is to be passed on as-is to preserve its full context
-	double NumberInput::getValue(State* state, unsigned int chunk_pos, double default_value){
+	float NumberInput::getValue(State* state, unsigned int chunk_pos, float default_value){
 		if (source){
 			return source->evaluate(state, chunk_pos);
 		} else {
